python-code/pyproject.toml
[project]
name = "computer-vision"
version = "0.1.0"
description = "Add your description here"
readme = "README.md"
requires-python = "==3.11.13"
dependencies = [
    "aiohttp>=3.13.5",
    "flask>=3.1.2",
    "ipykernel>=7.0.1",
    "jupyter>=1.1.1",
    "lap>=0.5.13",
    "livekit>=1.1.6",
    "livekit-api>=1.1.0",
    "matplotlib>=3.10.7",
    "mmcv==2.1.0",
    "mmdet>=3.3.0",
    "mmengine==0.10.6",
    "mmpose>=1.3.2",
    "numpy>=1.26.4",
    "opencv-python>=4.11.0.86",
    "openmim==0.3.9",
    "pandas>=2.3.3",
    "python-dotenv>=1.2.2",
    "scenedetect>=0.6.7.1",
    "scikit-learn>=1.7.2",
    "scipy>=1.16.2",
    "tensorflow>=2.20.0",
    "tensorflow-hub>=0.16.1",
    "torch>=2.9.0",
    "torchaudio>=2.9.0",
    "torchvision>=0.24.0",
    "tqdm>=4.67.1",
    "ultralytics==8.3.50",
]
